requirements.txt
Flask==0.10.1
docker-py==0.6.0
